backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
requests
pydantic
python-dotenv
flask
flask-cors
tensorflow
opencv-python
numpy
Pillow
